Moved NavBar to component directory. Moved Card to container directory
