+++ src/components/NavBar/style.scss
